+++ src/main/java/co/fineants/api/domain/member/config/MemberFactoryConfig.java
@@ -2,10 +2,13 @@ package co.fineants.api.domain.member.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.thymeleaf.spring6.SpringTemplateEngine;
 
 import co.fineants.api.domain.member.domain.factory.MemberFactory;
 import co.fineants.api.domain.member.domain.factory.MemberProfileFactory;
 import co.fineants.api.domain.member.domain.factory.MemberRoleFactory;
+import co.fineants.api.domain.member.domain.factory.MimeMessageFactory;
 import co.fineants.api.domain.member.service.RoleService;
 
 @Configuration
@@ -23,5 +26,10 @@ public class MemberFactoryConfig {
 	@Bean
 	public MemberProfileFactory memberProfileFactory() {
 		return new MemberProfileFactory();
+	}
+
+	@Bean
+	public MimeMessageFactory mimeMessageFactory(JavaMailSender mailSender, SpringTemplateEngine engine) {
+		return new MimeMessageFactory(mailSender, engine);
 	}
 }
